@@ -6,6 +6,7 @@ import Link from "next/link";
 import { createServerComponentClient } from "@supabase/auth-helpers-nextjs";
 import { cookies } from "next/headers";
 import { redirect } from "next/navigation";
+import { formatDistanceToNow } from "date-fns";
 
 export default async function SettingsPage() {
   const supabase = createServerComponentClient({ cookies });
@@ -15,28 +16,24 @@ export default async function SettingsPage() {
 
   const { data: profile } = await supabase
     .from('profiles')
-    .select('creator_status')
+    .select('role')
     .eq('id', session.user.id)
     .single();
 
   const { data: application } = await supabase
     .from('creator_applications')
-    .select('status, submitted_at')
+    .select('status, submitted_at, admin_notes')
     .eq('user_id', session.user.id)
     .single();
 
   const getStatusBadge = (status: string | null) => {
     switch (status) {
-      case 'pending_application':
+      case 'pending':
         return <Badge variant="secondary">Pending Review</Badge>;
-      case 'needs_onboarding':
-        return <Badge variant="default">Approved - Complete Onboarding</Badge>;
-      case 'active':
-        return <Badge variant="success">Active Creator</Badge>;
+      case 'approved':
+        return <Badge variant="default">Approved</Badge>;
       case 'denied':
-        return <Badge variant="destructive">Application Denied</Badge>;
-      case 'suspended':
-        return <Badge variant="destructive">Account Suspended</Badge>;
+        return <Badge variant="destructive">Denied</Badge>;
       default:
         return <Badge variant="outline">Not Applied</Badge>;
     }
@@ -44,16 +41,14 @@ export default async function SettingsPage() {
 
   const getStatusDescription = (status: string | null) => {
     switch (status) {
-      case 'pending_application':
+      case 'pending':
         return "Your application is being reviewed by our team. We'll notify you once a decision has been made.";
-      case 'needs_onboarding':
-        return "Congratulations! Your application has been approved. Complete the onboarding process to start creating content.";
-      case 'active':
-        return "You are an active creator on SubSpace. You can manage your content and subscriptions from your creator dashboard.";
+      case 'approved':
+        return "Congratulations! Your application has been approved. You can now start creating content.";
       case 'denied':
-        return "We're sorry, but your application was not approved at this time. You can apply again in the future.";
-      case 'suspended':
-        return "Your creator account has been suspended. Please contact support for more information.";
+        return application?.admin_notes 
+          ? `We're sorry, but your application was not approved at this time. Reason: ${application.admin_notes}`
+          : "We're sorry, but your application was not approved at this time. You can apply again in the future.";
       default:
         return "Join our creator program to start monetizing your content and building your community.";
     }
@@ -77,33 +72,57 @@ export default async function SettingsPage() {
             <CardHeader>
               <CardTitle>Creator Program</CardTitle>
               <CardDescription>
-                {getStatusDescription(profile?.creator_status)}
+                {getStatusDescription(application?.status)}
               </CardDescription>
             </CardHeader>
-            <CardContent className="space-y-4">
+            <CardContent className="space-y-6">
               <div className="flex items-center justify-between">
                 <div className="space-y-1">
                   <h3 className="text-lg font-medium">Application Status</h3>
-                  {getStatusBadge(profile?.creator_status)}
+                  {getStatusBadge(application?.status)}
                 </div>
-                {(!profile?.creator_status || profile.creator_status === 'denied') && (
+                {(!application || application.status === 'denied') && (
                   <Link href="/settings/creator/apply">
                     <Button>Apply Now</Button>
                   </Link>
                 )}
               </div>
 
-              {profile?.creator_status === 'needs_onboarding' && (
-                <div className="mt-4">
-                  <Link href="/settings/creator/onboarding">
-                    <Button>Complete Onboarding</Button>
-                  </Link>
+              {application && (
+                <div className="space-y-2">
+                  <div className="text-sm text-muted-foreground">
+                    Submitted {formatDistanceToNow(new Date(application.submitted_at), { addSuffix: true })}
+                  </div>
+                  {application.status === 'denied' && application.admin_notes && (
+                    <div className="mt-2 p-3 bg-destructive/10 rounded-md">
+                      <p className="text-sm text-destructive">
+                        {application.admin_notes}
+                      </p>
+                    </div>
+                  )}
                 </div>
               )}
 
-              {application && (
-                <div className="mt-4 text-sm text-muted-foreground">
-                  Submitted on: {new Date(application.submitted_at).toLocaleDateString()}
+              {profile?.role === 'creator' && (
+                <div className="mt-4 space-y-4">
+                  <div className="flex items-center gap-2">
+                    <Badge variant="outline">Creator Dashboard</Badge>
+                    <Link href="/creator/dashboard">
+                      <Button variant="link">View Dashboard</Button>
+                    </Link>
+                  </div>
+                  <div className="flex items-center gap-2">
+                    <Badge variant="outline">Content Management</Badge>
+                    <Link href="/creator/content">
+                      <Button variant="link">Manage Content</Button>
+                    </Link>
+                  </div>
+                  <div className="flex items-center gap-2">
+                    <Badge variant="outline">Subscription Settings</Badge>
+                    <Link href="/creator/subscriptions">
+                      <Button variant="link">Configure Subscriptions</Button>
+                    </Link>
+                  </div>
                 </div>
               )}
             </CardContent>
