@@ -2,20 +2,21 @@ import { createMiddlewareClient } from '@supabase/auth-helpers-nextjs';
 import { NextResponse } from 'next/server';
 import type { NextRequest } from 'next/server';
 
-export async function middleware(request: NextRequest) {
+export async function middleware(req: NextRequest) {
   const res = NextResponse.next();
-  const supabase = createMiddlewareClient({ req: request, res });
+  const supabase = createMiddlewareClient({ req, res });
 
   const {
     data: { session },
   } = await supabase.auth.getSession();
 
-  // Handle admin routes
-  if (request.nextUrl.pathname.startsWith('/admin')) {
+  // Check if the path starts with /admin
+  if (req.nextUrl.pathname.startsWith('/admin')) {
     if (!session) {
-      return NextResponse.redirect(new URL('/login', request.url));
+      return NextResponse.redirect(new URL('/login', req.url));
     }
 
+    // Check if user has admin role
     const { data: profile } = await supabase
       .from('profiles')
       .select('role')
@@ -23,28 +24,7 @@ export async function middleware(request: NextRequest) {
       .single();
 
     if (!profile || profile.role !== 'admin') {
-      return NextResponse.redirect(new URL('/', request.url));
-    }
-  }
-
-  // Handle creator onboarding
-  if (session) {
-    const { data: profile } = await supabase
-      .from('profiles')
-      .select('creator_status')
-      .eq('id', session.user.id)
-      .single();
-
-    // Redirect to onboarding if needed
-    if (profile?.creator_status === 'needs_onboarding' && 
-        !request.nextUrl.pathname.startsWith('/creator/onboarding')) {
-      return NextResponse.redirect(new URL('/creator/onboarding', request.url));
-    }
-
-    // Prevent accessing onboarding if not needed
-    if (profile?.creator_status !== 'needs_onboarding' && 
-        request.nextUrl.pathname.startsWith('/creator/onboarding')) {
-      return NextResponse.redirect(new URL('/creator/dashboard', request.url));
+      return NextResponse.redirect(new URL('/', req.url));
     }
   }
 
@@ -52,8 +32,5 @@ export async function middleware(request: NextRequest) {
 }
 
 export const config = {
-  matcher: [
-    '/admin/:path*',
-    '/creator/:path*',
-  ],
+  matcher: ['/admin/:path*'],
 }; 
